Add TabApplications component and necessary library
A new reusable TabApplications component has been added to the UI components, using the "@mui/lab" library for creating tab components. The component has been integrated into the main page and the "@mui/lab" was added to the package.json file as it is required for tab functionalities. This will contribute significantly to navigation within the application.