Add split button component and refactor button usage
Refactored the ButtonLarge component and replaced custom SuperButton with it. Also, a new component, ButtonSplit, has been added. The SuperButton has been removed for its redundancy. The ButtonLarge and ButtonSplit offer the same functionality but in a more organized and clean way improving the maintainability of the code. The changes have been carried through to the main index page.