Add Dialog and DialogConfirmEmail components
This commit introduces 2 new components namely 'Dialog' and 'DialogConfirmEmail'. Dialog component renders a customizable Dialog using MUI components including an optional title, text content, and buttons. The structure and behavior of the buttons are defined using the IDialogButtonProperties interface. DialogConfirmEmail on the other hand is a specific instance of the Dialog component which notifies users about serious consequences (like deletion of an account which is irreversible) and then takes confirmation from them. DialogConfirmEmail has been added to the main index page.