Add DialogGenerateInvite component and renames
Added a new DialogGenerateInvite component in the dialog folder. This component will display an invitation dialog which can be used for inviting other users to a certain event or group. A TextField has been included in the main Dialog component for allowing a user to input information, such as an invitation link. The DialogConfirmButton file and the Dialog have been renamed to make their purpose and usage more clear, the changes can be found in the index.tsx file of each component folder. Also, their import calls have been updated in the main index.tsx file.