Refactor styled components for messaging
This commit moves the styled-components for messaging and message field into their own module within and separate from the main component. This rearrangement makes the codebase more readable and maintainable, and it separates concerns better, with styling in one module and logic in another. Additionally, the Message and MessageField components have been moved to common from custom, to reflect that they are general UI components.