Add keyboard event listener for menu visibility control
The modification in the Page component includes the addition of a useEffect hook which is a keyboard event listener. This listens for the 'Escape' key press event to call the handleHideMenu function, acting as an additional method for the user to control the visibility of the Menu. This provides an alternative to mouse clicks and enhances user interactivity, improving the overall user experience.