Remove login, register, whoami mutations; Implement login via Telegram
The mutation files for login, register, and whoami were removed due to change in authentication method, shifting instead to logging in via Telegram. The login-via-telegram mutation was subsequently created. This approach provided a more secure option than logging in with email and password, thereby improving application security.

Also, the login page was modified to include a button for logging in via Telegram, making the new authentication method available to the users.

The local backend url was inserted into the .env_example, providing a fallback and easy setup for contributors if a remote backend url is not available.

Lastly, the type Account in the generated GraphQL schema was updated to reflect changes in the backend data model.