Add logic to generate group invite link
A 'createGroupInvite' mutation was added in order to generate a group invite link each time a private group is created. This is done by calling this mutation once a group is created, which then hits the server's 'createGroupInvite' endpoint. This logic was integrated into the 'create-group' page causing the page to fetch and display a group invite link after group creation. The groupInvite 'code' is appended to an invite link and set as the dialog's 'inputValue'. The invite link can then be copied to the clipboard. In addition, a 'group' query was added to fetch group data based on ID. Functional changes include disabling the 'Create Group' button and rendering 'Back' button after successful form submission.