Add invite page and logic for joining groups
This commit adds a new page, `invite.tsx`, that's responsible for handling the group joining process by accepting a group invite code. The page was built using a new `createGroupMemberMutation` which corresponds to a new mutation added in GraphQL schema.

On this page, two hooks are used to respectively redirect non-authenticated users to the login page and handle invite acceptance for authenticated users. If the acceptance is successful, the user will be redirected to either public or private group page, depending on the group type.

Apart from `invite.tsx`, a minor fix has been applied in constructing invite link where a slash was missing after the base URL.

Lastly, the changes also includes increasing the limit of fetching private and public groups from 10 to 100.

This commit provides end-to-end handling for group invitation - from sending an invitation to a user to successfully joining groups.