Update component references and enhance group functionality
Fixed references to a wrongly-named component and also modified the page for creating events to now include the group ID as a URL parameter. The "create event" page has been added and corresponding changes were made to the card for creating an event. This change was necessary to uniquely identify each group when working with groups, and allowing management of events within each group. Plus, this commit replaces the hard-coded texts in the UI with their actual data. This makes the UI more dynamic, user-friendly and aligns with application functionality.