Add group name availability check in create group page
Added a query to check if a group name is available in src/graphql/queries/is-group-name-available.graphql and used it in create-group.tsx. Upon creating a group, the system now checks if the inputted group name is available. If not, a snackbar notification is displayed with a message notifying that the group name already exists.

This feature is essential to prevent the duplication of group names, thus enhancing the user experience by providing real-time feedback on group name availability.