Refactor card-create-preference component; Add create-application page
The card-create-preference component was overhauled to make it more flexible by allowing the injection of any component as its child instead of having static TextField elements. This change enhance reusability of the card-create-preference component.

In addition, a new CreateApplication page has been added. This page presents a form for creating new applications, and it makes use of the newly refactored card-create-preference component by injecting TextField components into it. This page also includes relevant handlers for form submission and back navigation.  The changes were necessitated by the need to provide a functionality for creating new event applications in the application's frontend.