Make card-create-preference component more reusable; Add CreateApplication page
Refactored the card-create-preference component to enable the injection of any type of component as a child, enhancing its reusability. Added a new CreateApplication page which uses the refactored card-create-preference component. This new page presents a form for creating new applications and includes handlers for the form submission and back navigation. Changes were made to provide a means for creating new event applications on the frontend.